Extend rand_id_queue with push_front()
When a write back of modified data is triggered, the new write
transaction needs to be put in front of the expected AW queue for the
slave. The current method using push() and pop_front() only works if
the queue is max 1 entry long. With the optimizations of the ccu_fsm
added in the PROJ-239 branch, this situation may occur and causes tests
to fail.

The rand_id_queue class is extended and the push_front() method is added,
which allows an AW transaction to be put in the front of the AW queue.