wt_dcache_wbuffer: move user-bits deactivation
This fixes an issue with sv2v when DCACHE_USER_WIDTH is smaller than XLEN.
This is likely due to the bit-slicing operation overwriting other bits
from the packed struct when things aren't properly alligned.

Additionally setting the user bits to a constant-zero without additional condition
makes it so no sequential optimization is needed to remove these D-FF.
This is a non-issue for commercial tools but can be an issue for
open-source tools, which might not perform sequential optimizations at all.