Zero pad `wdata` tracer port based on `XLEN` (#2957)
Fix port connection width mismatch warning in the instruction tracer when using 32-bit version of the core. As the tracer is hardcoded to 64 bit operands and it could be useful to have the same trace format for both 64 and 32 bit cores, the most convenient solution is to pad the data when needed.

Co-authored-by: JeanRochCoulon <jean-roch.coulon@thalesgroup.com>